Fallback to github-actions bot user
If we are unable to find the login for the current token, fallback to `github-actions[bot]`. 

Failure to get the user for the github-runner token has been seen with GHES. This will allow PR comments to be created anyway - and if the token does belong the the github-actions app everything will work as it should.

If the token does not belong to the github-actions app and github still won't tell us who it belongs to, then applies will fail to find the comment. But there's not much we can do about that.