Isolate credentials inside container
Some credentials are required to be on disk in the home directory.
This directory is mounted by the runner, and some self-hosted runners don't clear them between jobs.
Instead write them to a directory inside the container and symlink to them, to stop them possibly leaking into another job.

Cleans up some unhelpful logs and use tree instead of ls